% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molstar.R
\name{molstarOutput}
\alias{molstarOutput}
\title{Create a Mol* Viewer for Shiny Applications}
\usage{
molstarOutput(
  outputId,
  pdbContents,
  defaultColor = "#D3D3D3",
  regionColors = NULL,
  width = "100\%",
  height = "400px",
  fill = FALSE
)
}
\arguments{
\item{outputId}{A character string specifying the output ID for the Mol* viewer.}

\item{pdbContents}{A character string containing the PDB file contents to be visualized.}

\item{defaultColor}{A character string specifying the default color for the molecular structure in hexadecimal format (e.g., "#D3D3D3"). Default is "#D3D3D3".}

\item{regionColors}{A data frame or matrix with columns 'selector' and 'color', where 'selector' specifies the region to color and 'color' is the hexadecimal color code. Default is NULL, meaning no region-specific coloring.}

\item{width}{A character string specifying the width of the viewer. Default is "100\%".}

\item{height}{A character string specifying the height of the viewer. Default is "400px".}

\item{fill}{A logical value indicating whether the viewer should fill the available space. Default is FALSE.}
}
\value{
A list of HTML elements and JavaScript code required to render the Mol* viewer in a Shiny application.
}
\description{
This function generates the necessary HTML and JavaScript to embed a Mol* viewer in a Shiny application. It allows for the visualization of molecular structures using PDB content and supports custom coloring of regions.
}
\examples{
\dontrun{
  library(shiny)
  ui <- fluidPage(
    molstarOutput("molstar_viewer", pdbContents = "PDB CONTENT HERE")
  )
  server <- function(input, output, session) {}
  shinyApp(ui, server)
}

}
